Restructure project, fix issues with installation
The header files are now restructured in the two following ways:
- They are moved to a separate `inc/` directory to keep source files and
header files separate
- They are renamed and placed under a separate `inc/libuimg/` directory
to both avoid having to prefix every header file with "libuimg_" and
also to make sure during installation that the user has access to
everything just by writing `#include <libuimg.h>` (the other header
files are imported in that "main" header file).

Furthermore, a call to `ldconfig` is added after the installation, to
make sure that ld's cache is updated to take the new library into
account, and an `uninstall` target is added to the Makefile, to make
removing installed libraries easier.